requirements.txt
psycopg2-binary
sshtunnel
PyYAML
python-dotenv
tabulate
openpyxl
